Refactor : 관리자 부분 파일 구조 수정
