+++ app/admin/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import { AdminPipelinePage } from '@/views/admin-pipeline';
+import { AdminPipelinePage } from '@/views/admin-pipeline/ui/AdminPipelinePage';
 
 export default function Page() {
   return (
