--- src/shared/ui/rich-text-editor/index.ts
@@ -1,1 +1,0 @@
-export { RichTextEditor } from './RichTextEditor';
