--- src/views/admin-pipeline/index.ts
@@ -1,1 +1,0 @@
-export { AdminPipelinePage } from './ui/AdminPipelinePage';
