--- src/views/content-management/index.ts
@@ -1,5 +1,0 @@
-export { ContentManagementPage } from './ui/ContentManagementPage';
-export { ContentDetailPage } from './ui/ContentDetailPage';
-export * from './model/labels';
-export * from './model/mockData';
-export * from './model/types';
